Struct ssssh::Handlers

pub struct Handlers<E, Pty = ()> where
    E: Into<Box<dyn StdError + Send + Sync + 'static>> + Send + 'static,  { /* fields omitted */ }

SSH callback handlers collections.

Implementations

impl<E, Pty> Handlers<E, Pty> where
E: Into<Box<dyn StdError + Send + Sync + 'static>> + Send + 'static,

pub fn new() -> Self

Construct new Handlers instance.

pub fn on_auth_none<H>(&mut self, handler: H) where
H: AuthNoneHandler<Error = E> + 'static,

Register None user authentication method handler.

If not registered, return none authentication failure.

Example

use ssssh::Handlers;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error>::new();
handlers.on_auth_none(|username| {
    async move {
        Ok(username == "bob")
    }.boxed()
});

pub fn on_auth_publickey<H>(&mut self, handler: H) where
H: AuthPublickeyHandler<Error = E> + 'static,

Register Publickey user authentication method handler.

If not registered, return publickey authentication failure.

Example

use ssssh::Handlers;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error>::new();
handlers.on_auth_publickey(|username, publickey: ssssh::PublicKey| {
    async move {
        let authorized_rsa_key =
        "AAAAB3NzaC1yc2EAAAADAQABAAABgQCsuW6XTH7zcwyQN9gKj3yVp9wg/4Hx5KL4YMXFBcjovr0KCA8NPvuYYn3WCyCO4zYoq4YrtjkS3XwRILjWo8Vx5zZcJL+zdGVLmQ5BNSWmvYAgcbpQrdftvk8y2SvMJHgK51g9cpumC8/D9yzOjNg1rlWQ0QZzDaUr0ugzQdL5KVXtTX3Mm3rjKhSy9coG7nJADv40R4tUiwJy0oorOn+E8y4lCdcNQnIxgME0WzgZ6NEJHU4s3cJY1OddWHRImunGLAJsSoAuHqpp8qtyuC8R+o+VcuqGLxXGCPoNNsy186dy7nGMCmGz+nJoNGR6jh+gHyHimGjqUticafo5NiY6J9uNjzh5HLg0B17iTR1iIDWDFyB3IRyNphnwEKl7OutNWvlk584b3USvTsVjBenNXKe181fE8s3hFs5B88NzXHoJuC+/L8/Y/tu24xckkt8ySCgRUHRJy9FOzmmpmaIeUZ9xB+IaQgn6Cue5tAzjeoa3wqyjlbV8lekK7DXlPOk=";
        Ok(username == "bob" && publickey.algorithm() == "ssh-rsa" && publickey.to_string() == authorized_rsa_key)
    }.boxed()
});

pub fn on_auth_publickey_signature_verified_after_accepted<H>(
&mut self,
handler: H
) where
H: AuthPublickeyHandler<Error = E> + 'static,

Register Publickey user authentication method handler. It handles When the verification of the message signature of the previously accepted public key is successful.

If not registered, return publickey authentication success.

Example

use ssssh::Handlers;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error>::new();
handlers.on_auth_publickey(|username, publickey: ssssh::PublicKey| {
    async move {
        let authorized_rsa_key =
        "AAAAB3NzaC1yc2EAAAADAQABAAABgQCsuW6XTH7zcwyQN9gKj3yVp9wg/4Hx5KL4YMXFBcjovr0KCA8NPvuYYn3WCyCO4zYoq4YrtjkS3XwRILjWo8Vx5zZcJL+zdGVLmQ5BNSWmvYAgcbpQrdftvk8y2SvMJHgK51g9cpumC8/D9yzOjNg1rlWQ0QZzDaUr0ugzQdL5KVXtTX3Mm3rjKhSy9coG7nJADv40R4tUiwJy0oorOn+E8y4lCdcNQnIxgME0WzgZ6NEJHU4s3cJY1OddWHRImunGLAJsSoAuHqpp8qtyuC8R+o+VcuqGLxXGCPoNNsy186dy7nGMCmGz+nJoNGR6jh+gHyHimGjqUticafo5NiY6J9uNjzh5HLg0B17iTR1iIDWDFyB3IRyNphnwEKl7OutNWvlk584b3USvTsVjBenNXKe181fE8s3hFs5B88NzXHoJuC+/L8/Y/tu24xckkt8ySCgRUHRJy9FOzmmpmaIeUZ9xB+IaQgn6Cue5tAzjeoa3wqyjlbV8lekK7DXlPOk=";
        Ok(username == "bob" && publickey.algorithm() == "ssh-rsa" && publickey.to_string() == authorized_rsa_key)
    }.boxed()
});
handlers.on_auth_publickey_signature_verified_after_accepted(|username, publickey:
ssssh::PublicKey| {
    async move {
        let authorized_rsa_key =
        "AAAAB3NzaC1yc2EAAAADAQABAAABgQCsuW6XTH7zcwyQN9gKj3yVp9wg/4Hx5KL4YMXFBcjovr0KCA8NPvuYYn3WCyCO4zYoq4YrtjkS3XwRILjWo8Vx5zZcJL+zdGVLmQ5BNSWmvYAgcbpQrdftvk8y2SvMJHgK51g9cpumC8/D9yzOjNg1rlWQ0QZzDaUr0ugzQdL5KVXtTX3Mm3rjKhSy9coG7nJADv40R4tUiwJy0oorOn+E8y4lCdcNQnIxgME0WzgZ6NEJHU4s3cJY1OddWHRImunGLAJsSoAuHqpp8qtyuC8R+o+VcuqGLxXGCPoNNsy186dy7nGMCmGz+nJoNGR6jh+gHyHimGjqUticafo5NiY6J9uNjzh5HLg0B17iTR1iIDWDFyB3IRyNphnwEKl7OutNWvlk584b3USvTsVjBenNXKe181fE8s3hFs5B88NzXHoJuC+/L8/Y/tu24xckkt8ySCgRUHRJy9FOzmmpmaIeUZ9xB+IaQgn6Cue5tAzjeoa3wqyjlbV8lekK7DXlPOk=";
        Ok(username == "bob" && publickey.algorithm() == "ssh-rsa" && publickey.to_string() == authorized_rsa_key)
    }.boxed()
});

pub fn on_auth_password<H>(&mut self, handler: H) where
H: AuthPasswordHandler<Error = E> + 'static,

Register Password user authentication method handler.

If not registered, return password authentication failure.

Example

use ssssh::Handlers;
use ssssh::PasswordResult;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error>::new();
handlers.on_auth_password(|username, password| {
    async move {
        Ok(if username == "bob" && password == "frosty-tricolor1-fabulous-unsent" {
            PasswordResult::Ok
        } else {
            PasswordResult::Failure
        })
    }.boxed()
});

pub fn on_auth_change_password<H>(&mut self, handler: H) where
H: AuthChangePasswordHandler<Error = E> + 'static,

Register Change Password user authentication method handler.

If not registered, return change password failure.

Example

use ssssh::Handlers;
use ssssh::PasswordResult;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error>::new();
handlers.on_auth_change_password(|username: String, oldpassword: String, newpassword:
String| {
    async move {
        let result = do_change_password(&username, &oldpassword, &newpassword);
        Ok(if result {
            PasswordResult::Ok
        } else {
            PasswordResult::Failure
        })
    }.boxed()
});

pub fn on_auth_hostbased<H>(&mut self, handler: H) where
H: AuthHostbasedHandler<Error = E> + 'static,

Register Hostbased user authentication method handler.

If not registered, return hostbased authentication failure.

Example

use ssssh::Handlers;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error>::new();
handlers.on_auth_hostbased(|username, hostname, publickey: ssssh::PublicKey| {
    async move {
        let authorized_rsa_key =
        "AAAAB3NzaC1yc2EAAAADAQABAAABgQCsuW6XTH7zcwyQN9gKj3yVp9wg/4Hx5KL4YMXFBcjovr0KCA8NPvuYYn3WCyCO4zYoq4YrtjkS3XwRILjWo8Vx5zZcJL+zdGVLmQ5BNSWmvYAgcbpQrdftvk8y2SvMJHgK51g9cpumC8/D9yzOjNg1rlWQ0QZzDaUr0ugzQdL5KVXtTX3Mm3rjKhSy9coG7nJADv40R4tUiwJy0oorOn+E8y4lCdcNQnIxgME0WzgZ6NEJHU4s3cJY1OddWHRImunGLAJsSoAuHqpp8qtyuC8R+o+VcuqGLxXGCPoNNsy186dy7nGMCmGz+nJoNGR6jh+gHyHimGjqUticafo5NiY6J9uNjzh5HLg0B17iTR1iIDWDFyB3IRyNphnwEKl7OutNWvlk584b3USvTsVjBenNXKe181fE8s3hFs5B88NzXHoJuC+/L8/Y/tu24xckkt8ySCgRUHRJy9FOzmmpmaIeUZ9xB+IaQgn6Cue5tAzjeoa3wqyjlbV8lekK7DXlPOk=";
        Ok(username == "bob" && hostname == "localhost." && publickey.algorithm() == "ssh-rsa" && publickey.to_string() == authorized_rsa_key)
    }.boxed()
});

pub fn on_channel_pty_request<H>(&mut self, handler: H) where
H: ChannelRequestPtyHandler<Pty, Error = E> + 'static,

Register Request pty handler.

If not registered, channel returns failure.

Example

use ssssh::Handlers;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error, Pty>::new();
handlers.on_channel_pty_request(|term: String, width, height, width_px, height_px, modes:
Vec<u8> | {
    async move {
        let pty: Pty = openpty(&term, width, height, width_px, height_px, &modes);
        Ok(pty)
    }.boxed()
});
struct Pty {
    // ...
}

pub fn on_channel_shell<H>(&mut self, handler: H) where
H: ChannelShellHandler<Pty, Error = E> + 'static,

Register Shell channel handler.

If not registered, channel returns failure.

Example

use ssssh::Handlers;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error>::new();
handlers.on_channel_shell(|mut ctx: ssssh::SessionContext<_>| {
    async move {
        let (stdin, stdout, stderr) = ctx.take_stdio().unwrap();
        let process = do_exec_shell(stdin, stdout, stderr);
        let exit_code = process.await;
        Ok(exit_code)
    }.boxed()
});

pub fn on_channel_exec<H>(&mut self, handler: H) where
H: ChannelExecHandler<Pty, Error = E> + 'static,

Register Exec channel handler.

If not registered, channel returns failure.

Example

use ssssh::Handlers;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error>::new();
handlers.on_channel_exec(|mut ctx: ssssh::SessionContext<_>, prog| {
    async move {
        let (stdin, stdout, stderr) = ctx.take_stdio().unwrap();
        let process = do_exec(prog, stdin, stdout, stderr);
        let exit_code = process.await;
        Ok(exit_code)
    }.boxed()
});

pub fn on_channel_direct_tcpip<H>(&mut self, handler: H) where
H: ChannelDirectTcpIpHandler<Error = E> + 'static,

Register Direct TCP/IP channel handler.

If not registered, channel returns failure.

Example

use ssssh::Handlers;
use futures::FutureExt as _;
let mut handlers = Handlers::<anyhow::Error>::new();
handlers.on_channel_direct_tcpip(|input, output| {
    async move {
        do_proxy(input, output).await;
        Ok(())
    }.boxed()
});